extract benchmark data
